done
& few fixes
