+++ CarLeasing.csproj
@@ -134,6 +134,7 @@
     <Compile Include="Controllers\AccountController.cs" />
     <Compile Include="Controllers\AdminPanelController.cs" />
     <Compile Include="Controllers\CarsPanelController.cs" />
+    <Compile Include="Controllers\ClientPanelController.cs" />
     <Compile Include="Controllers\IndexController.cs" />
     <Compile Include="Controllers\LoginController.cs" />
     <Compile Include="Controllers\ManufacturersController.cs" />
@@ -141,7 +142,6 @@
     <Compile Include="Controllers\OrdersPanelController.cs" />
     <Compile Include="Controllers\RegisterController.cs" />
     <Compile Include="Controllers\TiresPanelController.cs" />
-    <Compile Include="Controllers\UserPanelController.cs" />
     <Compile Include="Controllers\UsersPanelController.cs" />
     <Compile Include="LoginAuthenticationAttribute.cs" />
     <Compile Include="Global.asax.cs">
@@ -167,6 +167,7 @@
       <DependentUpon>CarLeasingModel.tt</DependentUpon>
     </Compile>
     <Compile Include="Models\EditCarAdminPanel.cs" />
+    <Compile Include="Models\EditClientPanel.cs" />
     <Compile Include="Models\firma.cs">
       <DependentUpon>CarLeasingModel.tt</DependentUpon>
     </Compile>
@@ -235,6 +236,7 @@
     <Content Include="App_Data\CarLeasing_log.ldf">
       <DependentUpon>CarLeasing.mdf</DependentUpon>
     </Content>
+    <Content Include="Content\AccountIndex.css" />
     <Content Include="Content\AdminPanel.css" />
     <Content Include="Content\bootstrap-grid.css" />
     <Content Include="Content\bootstrap-grid.min.css" />
@@ -246,6 +248,8 @@
     <Content Include="Content\CarsPanelDetails.css" />
     <Content Include="Content\CarsPanelEdit.css" />
     <Content Include="Content\CarsPanelIndex.css" />
+    <Content Include="Content\Images\Account\user.png" />
+    <Content Include="Content\Images\Account\userBg.jpg" />
     <Content Include="Content\Images\AdminPanel\cars.jpg" />
     <Content Include="Content\Images\AdminPanel\client.jpg" />
     <Content Include="Content\Images\AdminPanel\make.jpg" />
@@ -473,10 +477,6 @@
     <Content Include="Scripts\popper-utils.js.map" />
     <Content Include="Views\AdminPanel\Cars.cshtml" />
     <Content Include="Views\AdminPanel\Index.cshtml" />
-    <Content Include="Views\UserPanel\Delete.cshtml" />
-    <Content Include="Views\UserPanel\Details.cshtml" />
-    <Content Include="Views\UserPanel\Edit.cshtml" />
-    <Content Include="Views\UserPanel\Index.cshtml" />
     <Content Include="Views\Manufacturers\Create.cshtml" />
     <Content Include="Views\Manufacturers\Delete.cshtml" />
     <Content Include="Views\Manufacturers\Details.cshtml" />
@@ -502,6 +502,10 @@
     <Content Include="Views\OrdersPanel\Details.cshtml" />
     <Content Include="Views\OrdersPanel\Edit.cshtml" />
     <Content Include="Views\OrdersPanel\Index.cshtml" />
+    <Content Include="Views\ClientPanel\Delete.cshtml" />
+    <Content Include="Views\ClientPanel\Details.cshtml" />
+    <Content Include="Views\ClientPanel\Edit.cshtml" />
+    <Content Include="Views\ClientPanel\Index.cshtml" />
   </ItemGroup>
   <ItemGroup />
   <ItemGroup>
