+++ Views/Account/Index.cshtml
@@ -3,54 +3,25 @@
 @{
     ViewBag.Title = "Index";
 }
-
-<h2>Details</h2>
-
-<div>
-    <h4>uzytkownik</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.imie)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.imie)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.nazwisko)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.nazwisko)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.email)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.email)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.telefon)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.telefon)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.haslo)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.haslo)
-        </dd>
-    </dl>
+<link href="~/Content/AccountIndex.css" rel="stylesheet" />
+<div class="row mx-4">
+    <div class="col-12 pt-3 pr-5 bgHeader">
+        @Html.DisplayFor(model => model.imie) @Html.DisplayFor(model => model.nazwisko)
+    </div>
+    <div class="col-1"></div>
+    <div class="col-4 profile rounded">
+        <div class="userImgContainer">
+            <img src="~/Content/Images/Account/user.png" class="userImg rounded-circle img-fluid" />
+        </div>
+        <p class="text-center mt-2">
+            @Html.ActionLink("Edytuj", "Edit", new { id = Model.id_uzytkownik })
+        </p>
+        <div class="pt-4">
+            <ul>
+                <li> @Html.DisplayFor(model => model.email)</li>
+                <li> @Html.DisplayFor(model => model.telefon)</li>
+            </ul>
+        </div>
+    </div>
 </div>
-<p>
-    @Html.ActionLink("Edytuj", "Edit", new { id = Model.id_uzytkownik })
-</p>
+
