+++ Views/ClientPanel/Edit.cshtml
@@ -1,0 +1,64 @@
+﻿@model CarLeasing.Models.EditClientPanel
+
+@{
+    ViewBag.Title = "Edit";
+}
+<link href="~/Content/Breadcrumb.css" rel="stylesheet" />
+<div aria-label="breadcrumb">
+    <ol class="breadcrumb">
+        <li class="breadcrumb-item">@Html.ActionLink("Panel", "Index", "AdminPanel")</li>
+        <li class="breadcrumb-item">@Html.ActionLink("Lista", "Index")</li>
+        <li class="breadcrumb-item active" aria-current="page">Edytuj</li>
+    </ol>
+</div>
+@using (Html.BeginForm())
+{
+    @Html.AntiForgeryToken()
+
+    <div class="form-horizontal">
+        @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+        @Html.HiddenFor(model => model.id_zamowienie)
+
+        <div class="form-group">
+            @Html.LabelFor(model => model.okres_id_okres, "Długość wypożyczenia", htmlAttributes: new { @class = "control -label col-md-2" })
+            <div class="col-md-10">
+                @Html.DropDownList("okres_id_okres", null, htmlAttributes: new { @class = "form-control" })
+                @Html.ValidationMessageFor(model => model.okres_id_okres, "", new { @class = "text-danger" })
+            </div>
+        </div>
+
+        <div class="form-group">
+            @Html.LabelFor(model => model.platnosc_id_platnosc, "Płatność okresowa", htmlAttributes: new { @class = "control -label col-md-2" })
+            <div class="col-md-10">
+                @Html.DropDownList("platnosc_id_platnosc", null, htmlAttributes: new { @class = "form-control" })
+                @Html.ValidationMessageFor(model => model.platnosc_id_platnosc, "", new { @class = "text-danger" })
+            </div>
+        </div>
+
+        <div class="form-group">
+            @Html.LabelFor(model => model.dystans_id_dystans, "Dystans", htmlAttributes: new { @class = "control -label col-md-2" })
+            <div class="col-md-10">
+                @Html.DropDownList("dystans_id_dystans", null, htmlAttributes: new { @class = "form-control" })
+                @Html.ValidationMessageFor(model => model.dystans_id_dystans, "", new { @class = "text-danger" })
+            </div>
+        </div>
+
+        <div class="form-group">
+            @Html.LabelFor(model => model.kolor_id_kolor, "Kolor samochodu", htmlAttributes: new { @class = "control-label col-md-2" })
+            <div class="col-md-10">
+                @Html.DropDownList("kolor_id_kolor", null, htmlAttributes: new { @class = "form-control" })
+                @Html.ValidationMessageFor(model => model.kolor_id_kolor, "", new { @class = "text-danger" })
+            </div>
+        </div>
+
+        <div class="form-group">
+            <div class="col-md-offset-2 col-md-10">
+                <input type="submit" value="Zapisz" class="btn btn-default btnOrange" />
+            </div>
+        </div>
+    </div>
+}
+
+@section Scripts {
+    @Scripts.Render("~/bundles/jqueryval")
+}
