+++ Views/OrdersPanel/Details.cshtml
@@ -28,7 +28,7 @@
                             <td>@Html.DisplayFor(model => model.data_zlozenia)</td>
                         </tr>
                         <tr>
-                            <td>Okres</td>
+                            <td>Długość wypożyczenia</td>
                             <td>@Html.DisplayFor(model => model.okres.ilosc) mies.</td>
                         </tr>
                         <tr>
@@ -36,7 +36,7 @@
                             <td>@Html.DisplayFor(model => model.dystans.ilosc) km</td>
                         </tr>
                         <tr>
-                            <td>Rodzaj płatności</td>
+                            <td>Płatność okresowa</td>
                             <td>co @Html.DisplayFor(model => model.platnosc.ilosc) mies.</td>
                         </tr>
                         <tr>
