+++ Views/OrdersPanel/Index.cshtml
@@ -17,10 +17,10 @@
                 Dystans
             </th>
             <th>
-                Okres
+                Długość wypożyczenia
             </th>
             <th>
-                Rodzaj płatności
+                Płatność okresowa
             </th>
             <th>
                 Samochód
@@ -83,10 +83,10 @@
                 Dystans
             </th>
             <th>
-                Okres
+                Długość wypożyczenia
             </th>
             <th>
-                Rodzaj płatności
+                Płatność okresowa
             </th>
             <th>
                 Samochód
