--- Views/UserPanel/Edit.cshtml
@@ -1,55 +1,0 @@
-﻿@model CarLeasing.Models.zamowienie
-
-@{
-    ViewBag.Title = "Edit";
-}
-
-@using (Html.BeginForm())
-{
-    @Html.AntiForgeryToken()
-    
-<div class="form-horizontal">
-    <div class="form-group">
-        @Html.LabelFor(model => model.okres_id_okres, "Długość wypożyczenia", htmlAttributes: new { @class = "control-label col-md-2" })
-        <div class="col-md-10">
-            @Html.DropDownList("okres_id_okres", null, htmlAttributes: new { @class = "form-control" })
-        </div>
-    </div>
-
-    <div class="form-group">
-        @Html.LabelFor(model => model.platnosc_id_platnosc, "Płatność okresowa", htmlAttributes: new { @class = "control-label col-md-2" })
-        <div class="col-md-10">
-            @Html.DropDownList("platnosc_id_platnosc", null, htmlAttributes: new { @class = "form-control" })
-        </div>
-    </div>
-
-    <div class="form-group">
-        @Html.LabelFor(model => model.dystans_id_dystans, "Dystans", htmlAttributes: new { @class = "control-label col-md-2" })
-        <div class="col-md-10">
-            @Html.DropDownList("dystans_id_dystans", null, htmlAttributes: new { @class = "form-control" })
-        </div>
-    </div>
-
-    <div class="form-group">
-        @Html.LabelFor(model => model.koszt, "Koszt:", htmlAttributes: new { @class = "control-label col-md-2" })
-        <div class="col-md-10 price">
-            @ViewBag.cena
-        </div>
-        
-    </div>
-
-    <div class="form-group">
-        <div class="col-md-offset-2 col-md-10">
-            <input type="submit" value="Save" class="btn btn-primary" />
-        </div>
-    </div>
-</div>
-}
-
-<div>
-    @Html.ActionLink("Back to List", "Index")
-</div>
-
-@section Scripts {
-    @Scripts.Render("~/bundles/jqueryval")
-}
